+++ test/TestHelpers/TestFixtures.ts
@@ -177,6 +177,7 @@ import { PromoAsset, IPromoAsset } from 'Promo/Models/PromoAsset';
 import { PromoSize } from 'Promo/Models/PromoSize';
 import { AutomatedExperimentManager } from 'Ads/Managers/AutomatedExperimentManager';
 import { ClassDetectionApi } from 'Core/Native/ClassDetection';
+import { SdkDetectionInfo } from 'Core/Models/SdkDetectionInfo';
 
 const TestMediaID = 'beefcace-abcdefg-deadbeef';
 export class TestFixtures {
@@ -1009,7 +1010,8 @@ export class TestFixtures {
             MetaDataManager: new MetaDataManager(api),
             StorageBridge: new StorageBridge(api),
             ClientInfo: this.getClientInfo(platform),
-            Config: this.getCoreConfiguration()
+            Config: this.getCoreConfiguration(),
+            SdkDetectionInfo: new SdkDetectionInfo(platform, api)
         };
         if (platform === Platform.ANDROID) {
             core.DeviceInfo = new AndroidDeviceInfo(api);