+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Tue Mar 17 20:26:08 UTC 2020
+
+* Add isMadeWithUnity To New MWU Kafka Topic [#13768](https://github.com/Applifier/unity-ads-webview/pull/13768)
+* [Load V5] Move construction of CampaignManager to own method [#13685](https://github.com/Applifier/unity-ads-webview/pull/13685)
+* [LOAD V5] Improvements for Browser Test App [#13796](https://github.com/Applifier/unity-ads-webview/pull/13796)
+
 # Fri Mar 13 20:32:57 UTC 2020
 
 * AB Test for Auction XHR request on Android only [#13799](https://github.com/Applifier/unity-ads-webview/pull/13799)